refactor: code cleanup & remove unused code
