+++ src/main/java/ch/bfh/bti7081/s2020/orange/backend/data/MessageState.java
@@ -2,5 +2,5 @@ package ch.bfh.bti7081.s2020.orange.backend.data;
 
 public enum MessageState {
   UNREAD,
-  READ;
+  READ
 }
