+++ src/main/java/ch/bfh/bti7081/s2020/orange/backend/data/Mood.java
@@ -1,29 +1,29 @@
 package ch.bfh.bti7081.s2020.orange.backend.data;
 
 public enum Mood {
-    DEPRESSED("Deprimiert"),
-    ELATED("Begeistert"),
-    HAPPY("Glücklich"),
-    NEUTRAL("Normal"),
-    SAD("Traurig");
+  DEPRESSED("Deprimiert"),
+  ELATED("Begeistert"),
+  HAPPY("Glücklich"),
+  NEUTRAL("Normal"),
+  SAD("Traurig");
 
-    private final String label;
+  private final String label;
 
-    private Mood(String label) {
-        this.label = label;
+  Mood(String label) {
+    this.label = label;
+  }
+
+  public String getLabel() {
+    return label;
+  }
+
+  public static Mood valueOfLabel(String label) {
+    for (Mood e : values()) {
+      if (e.getLabel().equals(label)) {
+        return e;
+      }
     }
-
-    public String getLabel() {
-        return label;
-    }
-
-    public static Mood valueOfLabel(String label) {
-        for (Mood e : values()) {
-            if (e.getLabel().equals(label)) {
-                return e;
-            }
-        }
-        return null;
-    }
+    return null;
+  }
 }
