+++ src/main/java/ch/bfh/bti7081/s2020/orange/backend/service/HomeService.java
@@ -8,8 +8,5 @@ public class HomeService {
   public double calculate(long base, long power) {
     return Math.pow(base, power);
   }
-
-  public double getInitialValue() {
-    return 0;
-  }
+  
 }
