+++ src/main/java/ch/bfh/bti7081/s2020/orange/backend/service/MedicalSpecialistService.java
@@ -28,17 +28,9 @@ public class MedicalSpecialistService {
     return this.medicalSpecialistRepository.save(medicalSpecialist);
   }
 
-  public MedicalSpecialist getMedicalSpecialist(long id) {
-    return this.medicalSpecialistRepository.findById(id).get();
-  }
-
   public List<MedicalSpecialist> getAllMedicalSpecialist() {
 
     return this.medicalSpecialistRepository.findAll();
-  }
-
-  public void deleteMedicalSpecialist(long id) {
-    this.medicalSpecialistRepository.deleteById(id);
   }
 
 }
