+++ src/main/java/ch/bfh/bti7081/s2020/orange/backend/service/PatientService.java
@@ -15,7 +15,6 @@ import ch.bfh.bti7081.s2020.orange.backend.repositories.MoodDiaryRepository;
 import ch.bfh.bti7081.s2020.orange.backend.repositories.PatientRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -75,28 +74,6 @@ public class PatientService {
     p.setRole(Role.PATIENT);
 
     return this.patientRepository.save(p);
-  }
-
-  public Patient getPatient(long id) {
-
-    return this.patientRepository.findById(id).get();
-  }
-
-  public List<Patient> getPatientsByMedicalSpecialist(MedicalSpecialist ms) {
-    return this.patientRepository.findByMedicalSpecialist(ms);
-  }
-
-  public List<Patient> getPatientsWithoutMedicalSpecialist() {
-    return this.patientRepository.findByMedicalSpecialistNull();
-  }
-
-  public List<Patient> getAllPatients() {
-
-    return this.patientRepository.findAll();
-  }
-
-  public void deletePatient(long id) {
-    this.patientRepository.deleteById(id);
   }
 
 }
