+++ src/main/java/ch/bfh/bti7081/s2020/orange/ui/exceptions/AccessDeniedException.java
@@ -1,10 +1,11 @@
 package ch.bfh.bti7081.s2020.orange.ui.exceptions;
 
 public class AccessDeniedException extends RuntimeException {
-	public AccessDeniedException() {
-	}
 
-	public AccessDeniedException(String message) {
-		super(message);
-	}
+  public AccessDeniedException() {
+  }
+
+  public AccessDeniedException(String message) {
+    super(message);
+  }
 }
