+++ src/main/java/ch/bfh/bti7081/s2020/orange/ui/exceptions/UserAlreadyExistsException.java
@@ -1,10 +1,8 @@
 package ch.bfh.bti7081.s2020.orange.ui.exceptions;
 
 public class UserAlreadyExistsException extends RuntimeException {
-	public UserAlreadyExistsException() {
-	}
 
-	public UserAlreadyExistsException(String message) {
-		super(message);
-	}
+  public UserAlreadyExistsException(String message) {
+    super(message);
+  }
 }
