+++ src/main/java/ch/bfh/bti7081/s2020/orange/ui/utils/HasLogger.java
@@ -4,9 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * HasLogger is a feature interface that provides Logging capability for anyone
- * implementing it where logger needs to operate in serializable environment
- * without being static.
+ * HasLogger is a feature interface that provides Logging capability for anyone implementing it
+ * where logger needs to operate in serializable environment without being static.
  */
 public interface HasLogger {
