+++ src/main/java/ch/bfh/bti7081/s2020/orange/ui/views/chat/ChatPresenterImpl.java
@@ -26,7 +26,7 @@ public class ChatPresenterImpl implements ChatPresenter, ChatView.Observer {
 
   @PostConstruct
   public void init() {
-    messageService.getMessagesForChatId(chatId).subscribe(this::addMessageToView);
+    messageService.getMessages().subscribe(this::addMessageToView);
   }
 
   private void addMessageToView(Message message) {
